refactor: rename User -> concertUser
