+++ src/main/java/kr/hhplus/be/server/domain/concertUser/UserRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.concertUser;
+
+public interface UserRepository {
+    ConcertUser findById(long userId);
+}
